pyproject.toml
[project]
name = "case-study-matcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.31",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "langgraph-cli[inmem]>=0.3.8",
    "python-dotenv>=1.0.1",
]
